requirements.txt
rich>=9.10.0
eth-brownie>=1.13.0
